+++ src/main.ts
@@ -1,25 +1,22 @@
-import { setFailed, debug, info } from "@actions/core";
-import { env } from "process";
-import { parseInput } from "./input";
-import { OctokitGitHub } from "./github";
-import { Waiter } from "./wait";
+import { debug, info, setFailed } from '@actions/core';
+import { env } from 'process';
+import { OctokitGitHub } from './github';
+import { parseInput } from './input';
+import { Waiter } from './wait';
 
 async function run() {
   try {
     const input = parseInput(env);
     debug(
-      `Parsed inputs (w/o token): ${(({ githubToken, ...inputs }) =>
-        JSON.stringify(inputs))(input)}`,
+      `Parsed inputs (w/o token): ${(({ githubToken, ...inputs }) => JSON.stringify(inputs))(
+        input,
+      )}`,
     );
     const github = new OctokitGitHub(input.githubToken);
     debug(`Fetching workflows for ${input.owner}/${input.repo}...`);
     const workflows = await github.workflows(input.owner, input.repo);
-    debug(
-      `Found ${workflows.length} workflows in ${input.owner}/${input.repo}`,
-    );
-    const workflow_id = workflows.find(
-      (workflow) => workflow.name == input.workflowName,
-    )?.id;
+    debug(`Found ${workflows.length} workflows in ${input.owner}/${input.repo}`);
+    const workflow_id = workflows.find((workflow) => workflow.name == input.workflowName)?.id;
     if (workflow_id) {
       await new Waiter(workflow_id, github, input, info, debug).wait();
     } else {